makealbum: This program works again
I was not using this program so it was broken after the changes to
create makegallery. This fixes it up.

I also add a few comments to clarify fields/functions in this commit,
and expose the thumbnail/large image size default constants so they are
accessible by makealbum.

I update the album HTML building process to not include a gallery name
in the HTML if there is no name.