Fix issue where cropping thumbnails left empty pixels
I encountered a case where a cropped thumbnail ended up with size
300x100. The cropping succeeded in that only 100x100 of the image
remained, but on either side of this there were empty pixels. To fix
this I now call the equivalent of 'convert +repage' to reset the canvas
to include only what was cropped. See also
http://www.imagemagick.org/script/command-line-options.php#repage and
http://www.imagemagick.org/Usage/crop/#crop_page for how this works.

I also moved the AutoOrient calls to be the first action taken on
images, before any resizing or anything. I'm not sure this caused any
issues but it appeared incorrect.

Finally, a check calculating the offset to use when cropping was
incorrect. It was comparing height with height.